requirements.txt
pytest
pytest-selenium
python-dotenv
requests
schedule
selenium
